+++ packages/faas-cli-plugin-invoke/package.json
@@ -5,13 +5,7 @@
   "typings": "dist/index.d.ts",
   "dependencies": {
     "@midwayjs/fcli-command-core": "^0.1.12",
-    "@midwayjs/faas": "^0.1.5",
-    "@midwayjs/runtime-mock": "^0.1.12",
-    "@midwayjs/serverless-spec-builder": "^0.1.5",
-    "co": "^4.6.0",
-    "ts-node": "^8.5.2",
-    "urllib": "^2.34.1",
-    "websocket": "^1.0.30"
+    "@midwayjs/serverless-invoke": "^0.1.12"
   },
   "devDependencies": {
     "midway-bin": "1"
