--- packages/faas-cli-plugin-invoke/test/f.yml
@@ -1,13 +1,0 @@
-
-
-service: ginkgo-test
-
-provider:
-  name: ginkgo
-  runtime: node10tb
-
-functions:
-  index:
-    handler: index.handler
-    events:
-      - hsf: true
