--- packages/faas-cli-plugin-invoke/test/invoke.test.ts
@@ -1,10 +1,0 @@
-import { invoke } from '../src/invoke/main';
-// import * as assert from 'assert';
-describe('/test/invoke.test.ts', () => {
-  it('default invoke', async () => {
-    await invoke({
-      functionDir: __dirname,
-      functionName: 'index',
-    });
-  });
-});
