+++ packages/faas-cli/src/index.ts
@@ -1,10 +1,10 @@
 import { BaseCLI } from '@midwayjs/fcli-command-core';
 import { TestPlugin } from '@midwayjs/fcli-plugin-test';
-import { Invoke } from './invoke.plugin';
+import { InvokePlugin } from '@midwayjs/fcli-plugin-invoke';
 
 export class CLI extends BaseCLI {
   loadDefaultPlugin() {
-    this.core.addPlugin(Invoke);
+    this.core.addPlugin(InvokePlugin);
     this.core.addPlugin(TestPlugin);
   }
 }
