--- packages/faas-cli/src/invoke.plugin.ts
@@ -1,13 +1,0 @@
-import { InvokePlugin as BaseInvokePlugin } from '@midwayjs/fcli-plugin-invoke';
-import { runtimeEventMap } from '@midwayjs/invoke';
-
-export class Invoke extends BaseInvokePlugin {
-  getEventOptions(runtime: string, trigger: string) {
-    const runtimeMap = runtimeEventMap[runtime] || {};
-    return {
-      starter: runtimeMap.starter,
-      eventPath: runtimeMap.eventPath,
-      eventName: runtimeMap.eventName && runtimeMap.eventName[trigger],
-    };
-  }
-}
