+++ packages/serverless-invoke/package.json
@@ -1,14 +1,17 @@
 {
-  "name": "@midwayjs/invoke",
+  "name": "@midwayjs/serverless-invoke",
   "version": "0.1.12",
   "main": "dist/index",
   "typings": "dist/index.d.ts",
   "dependencies": {
-    "@midwayjs/fcli-plugin-invoke": "^0.1.12",
+    "@midwayjs/fcli-command-core": "^0.1.12",
     "@midwayjs/serverless-fc-starter": "^0.1.12",
     "@midwayjs/serverless-fc-trigger": "^0.1.12",
     "@midwayjs/serverless-scf-starter": "^0.1.12",
-    "@midwayjs/serverless-scf-trigger": "^0.1.12"
+    "@midwayjs/serverless-scf-trigger": "^0.1.12",
+    "ts-node": "^8.5.2",
+    "urllib": "^2.34.1",
+    "websocket": "^1.0.30"
   },
   "devDependencies": {
     "midway-bin": "1"
