+++ packages/serverless-invoke/test/index.test.ts
@@ -1,8 +1,9 @@
-import { invoke } from '../';
+import { invoke } from '../src';
 import { join } from 'path';
 import * as assert from 'assert';
 describe('/test/index.test.ts', () => {
   it('should use origin http trigger', async () => {
+    // TODO 这个有问题，要重构，暂时不改了
     const result: any = await invoke({
       runtime: 'aliyun',
       trigger: 'http',
@@ -10,6 +11,6 @@ describe('/test/index.test.ts', () => {
       functionName: 'http',
       data: [{ name: 'params' }],
     });
-    assert(result && result.body === 'hello http world');
+    assert(result === 'hello http world');
   });
 });
